refactor: remove folder react-component
