--- react-components/.husky/_/.gitignore
@@ -1,1 +1,0 @@
-*
