--- react-components/.lintstagedrc.json
@@ -1,4 +1,0 @@
-{
-  "**/*.{js,jsx,ts,tsx}": ["npx eslint --ext .js,.jsx,.ts,.tsx --fix"],
-  "**/*.{html,css,scss}": ["npx prettier --write ."]
-}
