--- react-components/src/components/Loading.tsx
@@ -1,10 +1,0 @@
-import loading from '../assets/hugging-grogu.gif';
-
-export default function Loading() {
-  return (
-    <div className="loading">
-      <img src={loading} alt="loading..." />
-      <div className="text-style">Loading...</div>
-    </div>
-  );
-}
