--- react-components/src/components/Logo.tsx
@@ -1,9 +1,0 @@
-export default function Logo() {
-  return (
-    <div className="logo">
-      <a className="logo-link text-style" href="/">
-        Star Wars
-      </a>
-    </div>
-  );
-}
