--- react-components/src/data/url.tsx
@@ -1,3 +1,0 @@
-const API_BASE_URL = 'https://swapi.dev/api/';
-
-export default API_BASE_URL;
