+++ src/assets/shooting-fire-din-djarin.gif
