+++ src/components/ErrorComponent.tsx
@@ -1,0 +1,4 @@
+export default function ErrorComponent() {
+  throw new Error(' This is a test error');
+  return <div>Error Boundary</div>;
+}
