+++ src/components/Header.tsx
@@ -1,0 +1,11 @@
+import Logo from './Logo';
+import ErrorButton from './ErrorButton';
+
+export default function Header() {
+  return (
+    <div className="header">
+      <Logo />
+      <ErrorButton />
+    </div>
+  );
+}
