+++ src/components/Logo.tsx
@@ -1,0 +1,9 @@
+export default function Logo() {
+  return (
+    <div className="logo">
+      <a className="logo-link text-style" href="/">
+        Star Wars
+      </a>
+    </div>
+  );
+}
