+++ src/data/url.tsx
@@ -1,0 +1,3 @@
+const API_BASE_URL = 'https://swapi.dev/api/';
+
+export default API_BASE_URL;
